server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq

# Vectorstore
chromadb
pinecone
langchain-pinecone

# Embeddings
sentence-transformers
langchain-huggingface
# langchain-google-generative-ai
# langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing, validation & Utilities
pydantic
requests

# Logging (optional but recommended)
# loguru

# File Uploads
# aiofiles

# Process
tqdm



# pip install -r requirements.txt